requirements.txt
langchain
langchain-ollama
flask
pymongo
python-dotenv
